Fixed the issue with the join slack button. Edited the subscribe to newsletter text. Added projects link to nav menu
