+++ src/components/shared/form/index.tsx
@@ -1,5 +1,6 @@
 export { CreateProjectForm } from './create-project';
 export { SignUpForm } from './sign-up';
+export { SignInForm } from './sign-in';
 export { ErrorMessage } from './error-msg';
 export { Form } from './form';
 export * from './controls';
