+++ src/components/shared/layout.tsx
@@ -26,6 +26,13 @@ const Layout: React.FC<LayoutProps> = ({ children, navLinks = [] }) => {
         link: true,
       },
       {
+        content: 'Blog',
+        href: '/blog',
+        title: `${siteMetadata.title} blog`,
+        requiresAuthentication: false,
+        link: true,
+      },
+      {
         content: 'Projects',
         href: '/app/projects',
         title: `${siteMetadata.title} projects`,
